+++ src/renderElement.spec.ts
@@ -5,146 +5,32 @@
  * ---------------------------------------------------------------------------------------------
  */
 import { create } from "./custom-element";
-import { renderToString } from "./renderToString";
+import { renderElement } from "./renderElement";
 
-describe("renderToString()", () => {
+describe("renderElement()", () => {
     it("returns outerHTML for single element", () => {
         const target = create("div", false).root();
 
-        const actual = renderToString(target);
+        const actual = renderElement(target, { indent: "" } as any);
 
         expect(actual).toBe("<div></div>");
     });
 
-    it("returns markup including shadow dom", () => {
+    it("returns element markup without shadow dom", () => {
         const target = create("div").shadowHtml(`<p>inside</p>`).root();
 
-        const actual = renderToString(target);
+        const actual = renderElement(target, { indent: "" } as any);
 
-        expect(actual).toMatchInlineSnapshot(`
-            "<div>
-                <template shadowroot=\\"open\\">
-                    <p>
-                        inside
-                    </p>
-                </template>
-            </div>"
-        `);
+        expect(actual).toMatchInlineSnapshot(`"<div></div>"`);
     });
 
-    it("returns markup including shadow dom with string attribute", () => {
-        const required = true;
-        const target = create("div")
-            .shadowHtml(
-                `<input type="text" id="name" name="name" ${required ? 'aria-required="true" required' : ""}
-        minlength="4" maxlength="8" size="10">`
-            )
-            .root();
-
-        const actual = renderToString(target);
-
-        expect(actual).toMatchInlineSnapshot(`
-            "<div>
-                <template shadowroot=\\"open\\">
-                    <input 
-                        aria-required=\\"true\\"
-                        id=\\"name\\"
-                        maxlength=\\"8\\"
-                        minlength=\\"4\\"
-                        name=\\"name\\"
-                        required=\\"\\"
-                        size=\\"10\\"
-                        type=\\"text\\" />
-                </template>
-            </div>"
-        `);
-    });
-
-    it("returns markup including shadow dom with void tag", () => {
-        const target = create("div").shadowHtml(`<br>`).root();
-
-        const actual = renderToString(target);
-
-        expect(actual).toMatchInlineSnapshot(`
-            "<div>
-                <template shadowroot=\\"open\\">
-                    <br />
-                </template>
-            </div>"
-        `);
-    });
-
-    it("returns markup including shadow dom with spaces", () => {
-        const target = create("div")
-            .shadowHtml(
-                `
-            <a
-                class="link link--display-inline-block link--bg"
-                href="expectedUrl"
-                target="_self"
-            >
-                Whatever
-            </a>
-            `
-            )
-            .root();
-
-        const actual = renderToString(target);
-
-        expect(actual).toMatchInlineSnapshot(`
-            "<div>
-                <template shadowroot=\\"open\\">
-                    <a 
-                        class=\\"link link--display-inline-block link--bg\\"
-                        href=\\"expectedUrl\\"
-                        target=\\"_self\\">
-                        
-                            Whatever
-                        
-                    </a>
-                </template>
-            </div>"
-        `);
-    });
-
-    it("returns markup including multiple levels of shadow dom", () => {
-        const target = create("div")
-            .shadowHtml(`<p>inside</p>`)
-            .shadowElements(create("div").shadowHtml(`<button>click me I'm inside of a second root</button>`).root())
-            .root();
-
-        const actual = renderToString(target);
-
-        expect(actual).toMatchInlineSnapshot(`
-            "<div>
-                <template shadowroot=\\"open\\">
-                    <p>
-                        inside
-                    </p>
-                    <div>
-                        <template shadowroot=\\"open\\">
-                            <button>
-                                click me I'm inside of a second root
-                            </button>
-                        </template>
-                    </div>
-                </template>
-            </div>"
-        `);
-    });
-
-    it("returns markup including light-dom children along shadow dom", () => {
+    it("returns markup including innerHTML w/o shadow root child", () => {
         const target = create("div").html(`<pre>I'm outside</pre>`).shadowHtml(`<p>inside</p>`).root();
 
-        const actual = renderToString(target);
+        const actual = renderElement(target, { indent: "", shadowDepths: Number.POSITIVE_INFINITY } as any);
 
         expect(actual).toMatchInlineSnapshot(`
             "<div>
-                <template shadowroot=\\"open\\">
-                    <p>
-                        inside
-                    </p>
-                </template>
                 <pre>
                     I'm outside
                 </pre>
@@ -155,7 +41,7 @@ describe("renderToString()", () => {
     it("returns only light-dom markup with shadowDepth option set to 0", () => {
         const target = create("div").html(`<pre>I'm outside</pre>`).shadowHtml(`<p>inside</p>`).root();
 
-        const actual = renderToString(target, { shadowDepth: 0 });
+        const actual = renderElement(target, { shadowDepth: 0, indent: "" } as any);
 
         expect(actual).toMatchInlineSnapshot(`
             "<div>
@@ -169,7 +55,7 @@ describe("renderToString()", () => {
     it("returns only light-dom markup with no shadowDepth set but dom set to 'light'", () => {
         const target = create("div").html(`<pre>I'm outside</pre>`).shadowHtml(`<p>inside</p>`).root();
 
-        const actual = renderToString(target, { shadow: false });
+        const actual = renderElement(target, { shadow: false, indent: "" } as any);
 
         expect(actual).toMatchInlineSnapshot(`
             "<div>
@@ -186,7 +72,7 @@ describe("renderToString()", () => {
             .shadowElements(create("div").shadowHtml(`<button>click me I'm inside of a second root</button>`).root())
             .root();
 
-        const actual = renderToString(target, { shadowDepth: 1 });
+        const actual = renderElement(target, { shadowDepth: 1, indent: "" } as any);
 
         expect(actual).toMatchInlineSnapshot(`
             "<div>
@@ -208,7 +94,7 @@ describe("filter tags", () => {
             .elements(create("main").html(`<button>click me I'm inside of a second element</button>`).root())
             .root();
 
-        const actual = renderToString(target, { filterTags: ["div"] });
+        const actual = renderElement(target, { filterTags: ["div"], indent: "" } as any);
 
         expect(actual).toMatchInlineSnapshot(`""`);
     });
@@ -219,7 +105,7 @@ describe("filter tags", () => {
             .elements(create("main").html(`<button>click me I'm inside of a second element</button>`).root())
             .root();
 
-        const actual = renderToString(target, { filterTags: ["button", "main"] });
+        const actual = renderElement(target, { filterTags: ["button", "main"], indent: "" } as any);
 
         expect(actual).toMatchInlineSnapshot(`
             "<div>
@@ -236,7 +122,7 @@ describe("filter tags", () => {
             .elements(create("main").html(`<button>click me I'm inside of a second element</button>`).root())
             .root();
 
-        const actual = renderToString(target, { filterTags: ["button"] });
+        const actual = renderElement(target, { filterTags: ["button"], indent: "" } as any);
 
         expect(actual).toMatchInlineSnapshot(`
             "<div>
@@ -258,17 +144,15 @@ describe("filter attributes", () => {
             )
             .root();
 
-        const actual = renderToString(target, { filterAttrs: ["id"] });
+        const actual = renderElement(target, { filterAttrs: ["id"], indent: "" } as any);
 
         expect(actual).toMatchInlineSnapshot(`
             "<div>
-                <p 
-                    >
+                <p>
                     inside the element
                 </p>
                 <main>
-                    <button 
-                        >
+                    <button>
                         click me I'm inside of a second element
                     </button>
                 </main>
@@ -286,17 +170,15 @@ describe("filter attributes", () => {
             )
             .root();
 
-        const actual = renderToString(target, { filterAttrs: ["id"] });
+        const actual = renderElement(target, { filterAttrs: ["id"], indent: "" } as any);
 
         expect(actual).toMatchInlineSnapshot(`
             "<div>
-                <p 
-                    align=\\"center\\">
+                <p align=\\"center\\">
                     inside the element
                 </p>
                 <main>
-                    <button 
-                        class=\\"favorite\\">
+                    <button class=\\"favorite\\">
                         click me I'm inside of a second element
                     </button>
                 </main>